refactor(minecraft): ブリッジ／メモリツールとその spec を minecraft へ移譲する
