+++ packages/mcp/package.json
@@ -8,15 +8,11 @@
 		"./http-server": "./src/http-server.ts",
 		"./memory-cache": "./src/memory-cache.ts",
 		"./lru-cache": "./src/lru-cache.ts",
-		"./memory-helpers": "./src/memory-helpers.ts",
 		"./emotion": "./src/emotion.ts",
 		"./test-helpers": "./src/test-helpers.ts",
 		"./tool-metrics": "./src/tool-metrics.ts",
 		"./tools/discord": "./src/tools/discord.ts",
-		"./tools/event-helpers": "./src/tools/event-helpers.ts",
 		"./tools/mc-bridge-discord": "./src/tools/mc-bridge-discord.ts",
-		"./tools/mc-bridge-minecraft": "./src/tools/mc-bridge-minecraft.ts",
-		"./tools/mc-memory": "./src/tools/mc-memory.ts",
 		"./tools/memory": "./src/tools/memory.ts",
 		"./tools/result": "./src/tools/result.ts",
 		"./tools/schedule": "./src/tools/schedule.ts",
