+++ packages/minecraft/package.json
@@ -3,6 +3,10 @@
 	"private": true,
 	"exports": {
 		"./constants": "./src/constants.ts",
+		"./mc-bridge-tools": "./src/mc-bridge-tools.ts",
+		"./mc-memory": "./src/mc-memory.ts",
+		"./event-helpers": "./src/event-helpers.ts",
+		"./memory-helpers": "./src/memory-helpers.ts",
 		"./server": "./src/server.ts",
 		"./mc-bridge-server": "./src/mc-bridge-server.ts",
 		"./mcp-tools": "./src/mcp-tools.ts",
