+++ packages/minecraft/src/memory-helpers.ts
