+++ spec/minecraft/actions/shared.spec.ts
