+++ spec/minecraft/actions/smelt.spec.ts
