+++ spec/minecraft/actions/survival.spec.ts
