+++ spec/minecraft/auto-notifier.spec.ts
