+++ spec/minecraft/bot-context.spec.ts
