+++ spec/minecraft/bot-queries.spec.ts
