+++ spec/minecraft/entrypoints.spec.ts
