+++ spec/minecraft/exploration.spec.ts
