+++ spec/minecraft/mc-bridge-format-commands.spec.ts
@@ -1,7 +1,7 @@
 import { describe, expect, test } from "bun:test";
 
-import type { ErrorEvent, ParsedEvent } from "@vicissitude/mcp/tools/event-helpers";
-import { formatCommands } from "@vicissitude/mcp/tools/mc-bridge-minecraft";
+import type { ErrorEvent, ParsedEvent } from "@vicissitude/minecraft/event-helpers";
+import { formatCommands } from "@vicissitude/minecraft/mc-bridge-tools";
 
 /** デフォルト値付きの ParsedEvent ファクトリ。テストごとに必要なフィールドだけ上書きする */
 function createEvent(overrides: Partial<ParsedEvent> = {}): ParsedEvent {
