+++ spec/minecraft/queries.spec.ts
