+++ spec/minecraft/stub-logger.ts
