+++ spec/minecraft/stuck-recovery.spec.ts
